refactor: extract OAuth and MCP logic into service modules
Move auth-code store, client registry, PKCE verification, JWT issuance,
and MCP tool definitions/dispatch out of the route files and into
src/services/oauth.ts and src/services/mcp.ts. Route files are now thin
HTTP adapters. Also removes the duplicated jwtSecret() from
require-bearer.ts — it now imports from the OAuth service.
